api-service/requirements.txt
# Web framework
fastapi==0.105.0
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Authentication
pyjwt==2.8.0
cryptography==41.0.8
python-jose[cryptography]==3.3.0

# Database
duckdb==0.9.2
boto3==1.34.0

# Rate limiting
slowapi==0.1.9

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
